Fix ghost ping by edit bug & update dependencies
The ghost ping by edit message was being sent even if the message contents were the same, suggesting some other message property could've changed which is not the embeds.
Anyway, if the contents are the same it means the event was not triggered by a user editing their message so can ignore all such events.